+++ CHANGELOG.md
@@ -18,6 +18,7 @@ with the exception that 0.x versions can break between minor versions.
 - Add support for Oniguruma's absent repeater `(?~abc)` (#233)
 - Add support for ignoring empty matches (#240)
 - Add support for treating unnamed capture groups as non-capturing when named groups exist (#241)
+- Add experimental seek optimization to only invoke the VM at candidate positions where a match could occur (#246)
 ### Changed
 - `RegexBuilder` can now build multiple patterns with the same options (#213)
 - Parsing of capture group names is now more lenient (#241)